ECONOMY MADNESS? THIS IS … FIXED

+++ Essentials/src/com/earth2me/essentials/ISettings.java
@@ -79,7 +79,7 @@ public interface ISettings extends IConf
 
 	int getSpawnMobLimit();
 
-	int getStartingBalance();
+	BigDecimal getStartingBalance();
 
 	double getTeleportCooldown();
